requirements.txt
langchain
dspy
pinecone-client
sentence-transformers
transformers
streamlit
peft
torch
datasets
python-dotenv
